ZoroCEO token - ZRC
Zoro CEO is a community project that offers daily and monthly prizes to lucky ZRC holders. Thanks to the development team's various burning mechanisms, ZRC is constantly being burned, making it increasingly scarce with the support of the community. Together, the development team aims to make ZRC the best meme coin of 2023.